requirements.txt
opencv-python-headless
streamlit
numpy
Pillow
